third_party/blink/web_tests/external/wpt/css/cssom/shorthand-values-expected.txt

This is a testharness.js-based test.
[FAIL] The serialization of list-style-type: circle; list-style-position: inside; list-style-image: none; should be canonical.
  assert_equals: expected "list-style: inside circle;" but got "list-style: inside none circle;"
Harness: the test ran to completion.
